VMC/sandbox/requirements.txt
bell-avr-libraries[mqtt]==0.1.12
loguru==0.6.0
pillow
opencv-python-headless
scipy
